zobov.py - fix bug that moves all files if handle is empty
An empty value of `handle` in `parameters/params.py` can cause
zobov to shift all files from the Revolver/ main directory into a
subsubdirectory and then crash due to recursiveness.

This commit fixes the bug by setting a default handle "myhandle"
if handle is an empty string.

Side effect: this commit also solves a bug that I haven't posted: if
`output_folder` does not have a trailing slash, then the zobov
directory handling strategies are again unlikely to do what is
expected by the user.  It's easy to use a python library function to
add a trailing slash in an OS-independent way, so that is done
in this commit too.